Finalizados los value objects

+++ domain/src/main/java/co/com/sofkau/icecreamshop/sale/values/Price.java
@@ -2,18 +2,17 @@ package co.com.sofkau.icecreamshop.sale.values;
 
 import co.com.sofka.domain.generic.ValueObject;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
-public class Price implements ValueObject<BigDecimal> {
-    private final BigDecimal value;
+public class Price implements ValueObject<Integer> {
+    private final Integer value;
 
-    public Price(BigDecimal value) {
+    public Price(Integer value) {
         this.value = Objects.requireNonNull(value);
     }
 
     @Override
-    public BigDecimal value() {
+    public Integer value() {
         return this.value;
     }
 }
